Move find_spans to utils
